move walker/wrapper.hh -> walker.hh
